--- src/main/java/edu/kis/powp/jobs2d/command/InterfaceAdapter.java
@@ -1,42 +1,0 @@
-package edu.kis.powp.jobs2d.command;
-
-import com.google.gson.*;
-
-import java.lang.reflect.Type;
-
-//TODO figure out naming and location
-
-
-public class InterfaceAdapter implements JsonSerializer, JsonDeserializer {
-
-    private static final String CLASSNAME = "CLASSNAME";
-    private static final String DATA = "DATA";
-    private static final String PACKAGE = "edu.kis.powp.jobs2d.command.";
-
-    public Object deserialize(JsonElement jsonElement, Type type,
-                              JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
-
-        JsonObject jsonObject = jsonElement.getAsJsonObject();
-        JsonPrimitive prim = (JsonPrimitive) jsonObject.get(CLASSNAME);
-        String className = PACKAGE.concat(prim.getAsString());
-        Class klass = getObjectClass(className);
-        return jsonDeserializationContext.deserialize(jsonObject.get(DATA), klass);
-    }
-
-    public JsonElement serialize(Object jsonElement, Type type, JsonSerializationContext jsonSerializationContext) {
-        JsonObject jsonObject = new JsonObject();
-        jsonObject.addProperty(CLASSNAME, jsonElement.getClass().getName());
-        jsonObject.add(DATA, jsonSerializationContext.serialize(jsonElement));
-        return jsonObject;
-    }
-
-    public Class getObjectClass(String className) {
-        try {
-            return Class.forName(className);
-
-        } catch (ClassNotFoundException e) {
-            //e.printStackTrace();
-            throw new JsonParseException(e.getMessage());
-        }
-    }
-}
